refactor: util 내부 파일들을 utils.js에 통합, Lotto가 할 수 있는 기능은 Lotto내부에서 메서드로 선언
