--- src/util/sort.js
@@ -1,3 +1,0 @@
-export const getAscendingSortedNumbers = (numbers) => {
-  return numbers.sort((a, b) => a - b);
-};
